refactor: add SellNativeWarningBanner non-pure
